+++ test/plots/index.ts
@@ -8,6 +8,7 @@ export * from "./aapl-monthly.js";
 export * from "./aapl-volume-rect.js";
 export * from "./aapl-volume.js";
 export * from "./anscombe-quartet.js";
+export * from "./arc.js";
 export * from "./armadillo.js";
 export * from "./aspectRatio.js";
 export * from "./athletes-bins-colors.js";
@@ -49,7 +50,6 @@ export * from "./cars-parcoords.js";
 export * from "./channel-domain.js";
 export * from "./clamp.js";
 export * from "./collapsed-histogram.js";
-export * from "./collatz.js";
 export * from "./color-piecewise.js";
 export * from "./country-centroids.js";
 export * from "./covid-ihme-projected-deaths.js";
@@ -157,7 +157,6 @@ export * from "./metro-unemployment-ridgeline.js";
 export * from "./metro-unemployment-slope.js";
 export * from "./metro-unemployment-stroke.js";
 export * from "./metro-unemployment.js";
-export * from "./miserables.js";
 export * from "./moby-dick-faceted.js";
 export * from "./moby-dick-letter-frequency.js";
 export * from "./moby-dick-letter-pairs.js";
